# swagger.yaml
swagger: "2.0"
info:
    version: "1.0.0"
    title: "Kessoku API"
paths:
    /account/login:
        post:
            tags:
                - "Authentication"
            summary: "User login"
            description: "This route allows users to login by providing their username, password and remember-me option."
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                - in: "body"
                  name: "body"
                  description: "User login information"
                  required: true
                  schema:
                      type: "object"
                      properties:
                          username:
                              type: "string"
                          password:
                              type: "string"
                          remember:
                              type: "boolean"
                      required:
                          - "username"
                          - "password"
                          - "remember"
            responses:
                "200":
                    description: "Login successful"
                    schema:
                        type: "object"
                        properties:
                            username:
                                type: "string"
                "401":
                    description: "Invalid input or Invalid credentials"
                "500":
                    description: "Internal server error"
    /account/register:
        post:
            tags:
                - "Authentication"
            summary: "User registration"
            description: "This route allows new users to register by providing a username and password."
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                - in: "body"
                  name: "body"
                  description: "User registration information"
                  required: true
                  schema:
                      type: "object"
                      properties:
                          username:
                              type: "string"
                              pattern: "^[A-Za-z][A-Za-z0-9_]{0,31}$"
                              description: "Between 1 and 32 characters long and must not start with a number or special characters (only letters at the beginning), and the rest of the characters can include letters, numbers, or special characters."
                          password:
                              type: "string"
                              pattern: "^[A-Za-z][A-Za-z0-9_]{5,31}$"
                              description: "Between 6 and 32 characters long and must not start with a number or special characters (only letters at the beginning), and the rest of the characters can include letters, numbers, or special characters."
                      required:
                          - "username"
                          - "password"
            responses:
                "201":
                    description: "Registration successful"
                    schema:
                        type: "object"
                        properties:
                            username:
                                type: "string"
                "400":
                    description: "Invalid input"
                    schema:
                        type: "array"
                        items:
                            type: "object"
                            properties:
                                path:
                                    type: "array"
                                    items:
                                        type: "string"
                                message:
                                    type: "string"
                "409":
                    description: "Username already taken"
                "500":
                    description: "Internal server error"
    /account/logout:
        get:
            tags:
                - "Authentication"
            summary: "User logout"
            description: "This route allows users to logout from the system."
            produces:
                - "application/json"
            responses:
                "200":
                    description: "Successfully logged out"
                "500":
                    description: "Internal server error"
    /api/account/verify:
        post:
            tags:
                - "Authentication"
            summary: "Starts the user verification process"
            description: "This route is used to start the user verification process. It requires the user to be logged in and not already verified."
            parameters:
                - in: "body"
                  name: "body"
                  description: "Email of the user to be verified"
                  required: true
                  schema:
                      type: "object"
                      properties:
                          email:
                              type: "string"
            responses:
                "200":
                    description: "Verification process started successfully"
                "400":
                    description: "Bad request, email is required or user is already in verification process"
                "401":
                    description: "Unauthorized, user is not logged in"
                "409":
                    description: "Conflict, user is already verified"
                "500":
                    description: "Internal server error"
    /api/account/verify/{token}:
        get:
            tags:
                - "Authentication"
            summary: "Completes the user verification process"
            description: "This route is used to complete the user verification process. It requires a valid JWT token."
            parameters:
                - in: "path"
                  name: "token"
                  description: "JWT token for verification"
                  required: true
                  type: "string"
            responses:
                "200":
                    description: "User verified successfully"
                "400":
                    description: "Bad request, invalid token"
                "404":
                    description: "Not found, user not found"
    /api/thread/comment:
        post:
            tags:
                - "Comments"
            summary: "Create a new comment"
            description: "This route allows a logged-in user to create a new comment on a post."
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                - in: "body"
                  name: "body"
                  description: "Comment information"
                  required: true
                  schema:
                      type: "object"
                      properties:
                          content:
                              type: "string"
                          postId:
                              type: "string"
                      required:
                          - "content"
                          - "postId"
            responses:
                "200":
                    description: "Comment created successfully"
                    schema:
                        type: "object"
                        properties:
                            content:
                                type: "string"
                            authorId:
                                type: "string"
                            postId:
                                type: "string"
                "401":
                    description: "Invalid input or Unauthorized"
                "500":
                    description: "Internal server error"
    /api/thread/comment/post/{id}:
        get:
            tags:
                - "Comments"
            summary: "Get comments by post ID"
            description: "This route allows users to retrieve all comments associated with a specific post ID."
            produces:
                - "application/json"
            parameters:
                - in: "path"
                  name: "id"
                  description: "Post ID"
                  required: true
                  type: "string"
            responses:
                "200":
                    description: "Comments retrieved successfully"
                    schema:
                        type: "array"
                        items:
                            type: "object"
                            properties:
                                content:
                                    type: "string"
                                authorId:
                                    type: "string"
                                postId:
                                    type: "string"
                "401":
                    description: "Invalid input"
                "500":
                    description: "Internal server error"
    /api/thread/comment/{id}:
        get:
            tags:
                - "Comments"
            summary: "Get a comment by ID"
            description: "This route allows users to retrieve a comment by its ID."
            produces:
                - "application/json"
            parameters:
                - in: "path"
                  name: "id"
                  description: "Comment ID"
                  required: true
                  type: "string"
            responses:
                "200":
                    description: "Comment retrieved successfully"
                    schema:
                        type: "object"
                        properties:
                            content:
                                type: "string"
                            authorId:
                                type: "string"
                            postId:
                                type: "string"
                "401":
                    description: "Invalid input"
                "500":
                    description: "Internal server error"
    /api/thread/post:
        post:
            tags:
                - "Posts"
            summary: "Create a new post"
            description: "This route allows a logged-in user to create a new post."
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                - in: "body"
                  name: "body"
                  description: "Post information"
                  required: true
                  schema:
                      type: "object"
                      properties:
                          title:
                              type: "string"
                          content:
                              type: "string"
                          topic:
                              type: "string"
                      required:
                          - "title"
                          - "content"
                          - "topic"
            responses:
                "200":
                    description: "Post created successfully"
                    schema:
                        type: "object"
                        properties:
                            title:
                                type: "string"
                            content:
                                type: "string"
                            authorId:
                                type: "string"
                            topicId:
                                type: "string"
                "401":
                    description: "Invalid input, topic not found or Unauthorized"
                "500":
                    description: "Internal server error"
    /api/thread/post/{id}:
        get:
            tags:
                - "Posts"
            summary: "Get a post by ID"
            description: "This route allows users to retrieve a post by its ID."
            produces:
                - "application/json"
            parameters:
                - in: "path"
                  name: "id"
                  description: "Post ID"
                  required: true
                  type: "string"
            responses:
                "200":
                    description: "Post retrieved successfully"
                    schema:
                        type: "object"
                        properties:
                            title:
                                type: "string"
                            content:
                                type: "string"
                            authorId:
                                type: "string"
                            topicId:
                                type: "string"
                "401":
                    description: "Invalid input"
                "500":
                    description: "Internal server error"